+++ tests.sh
@@ -1,0 +1,1 @@
+node qdbp examples/example1.qdbp examples/example1.sexpr